Use uint64 rather than int for r-tree record ID
By using a uint64, we open up the possibility do do things other than
just storing an index to an array with the record id. E.g. we could
treat the uint64 as two uint32s, and store two different indexes. This
would be less clean with an int, because we can't be 100% guarenteed
that we have 64 bits to play with.

+++ geom/alg_intersection.go
@@ -21,7 +21,7 @@ func intersectionOfLines(lines1, lines2 []line) (MultiPoint, MultiLineString) {
 	for i, ln := range lines1 {
 		bulk[i] = rtree.BulkItem{
 			Box:      toBox(ln.envelope()),
-			RecordID: i,
+			RecordID: uint64(i),
 		}
 	}
 	tree := rtree.BulkLoad(bulk)
@@ -31,7 +31,7 @@ func intersectionOfLines(lines1, lines2 []line) (MultiPoint, MultiLineString) {
 	seen := make(map[XY]bool)
 
 	for j := range lines2 {
-		tree.Search(toBox(lines2[j].envelope()), func(i int) error {
+		tree.Search(toBox(lines2[j].envelope()), func(i uint64) error {
 			inter := lines2[j].intersectLine(lines1[i])
 			if inter.empty {
 				return nil
